Move useBen and useEmitProxy to composables.
